+++ app/src/main/java/com/example/compose_magazin/data/mappers/Mappers.kt
@@ -1,13 +1,13 @@
 package com.example.compose_magazin.data.mappers
 
-import com.example.compose_magazin.data.models.PetCategoryData
-import com.example.compose_magazin.data.models.PetProductData
-import com.example.compose_magazin.data.models.PetTagData
+import com.example.compose_magazin.data.models.PetCategoryNetworkEntity
+import com.example.compose_magazin.data.models.PetProductNetworkEntity
+import com.example.compose_magazin.data.models.PetTagNetworkEntity
 import com.example.compose_magazin.domain.entity.PetCategory
 import com.example.compose_magazin.domain.entity.PetProduct
 import com.example.compose_magazin.domain.entity.PetTag
 
-fun petProductDataToPetProduct(data: PetProductData): PetProduct =
+fun petProductDataToPetProduct(data: PetProductNetworkEntity): PetProduct =
     PetProduct(
         id = data.id,
         name = data.name,
@@ -17,13 +17,13 @@ fun petProductDataToPetProduct(data: PetProductData): PetProduct =
         status = data.status
     )
 
-fun petTagDataToPetTag(data: PetTagData): PetTag =
+fun petTagDataToPetTag(data: PetTagNetworkEntity): PetTag =
     PetTag(
         id = data.id,
         name = data.name
     )
 
-fun petCategoryDataToPetCategory(data: PetCategoryData): PetCategory =
+fun petCategoryDataToPetCategory(data: PetCategoryNetworkEntity): PetCategory =
     PetCategory(
         id = data.id,
         name = data.name
