+++ app/src/main/java/com/example/compose_magazin/data/models/PetCategoryNetworkEntity.kt
@@ -3,7 +3,7 @@ package com.example.compose_magazin.data.models
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class PetCategoryData(
+data class PetCategoryNetworkEntity(
     val id: Long,
     val name: String? = ""
 )
