--- app/src/main/java/com/example/compose_magazin/data/models/PetProductData.kt
@@ -1,13 +1,0 @@
-package com.example.compose_magazin.data.models
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class PetProductData(
-    val id: Long,
-    val name: String? = "unknown",
-    val category: PetCategoryData? = PetCategoryData(-1, ""),
-    val tags: List<PetTagData>? = null,
-    val photoUrls: List<String?>? = listOf(),
-    val status: String?
-)
