--- app/src/main/java/com/example/compose_magazin/data/models/PetTagData.kt
@@ -1,9 +1,0 @@
-package com.example.compose_magazin.data.models
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class PetTagData(
-    val id: Long,
-    val name: String? = null
-)
