--- app/src/main/java/com/example/compose_magazin/dataAccess/api/RetrofitClient.kt
@@ -1,35 +1,0 @@
-package com.example.compose_magazin.dataAccess.api
-
-import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
-import kotlinx.serialization.ExperimentalSerializationApi
-import kotlinx.serialization.json.Json
-import okhttp3.MediaType.Companion.toMediaType
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-
-object RetrofitClient {
-    private const val BASE_URL = "https://petstore.swagger.io/v2/"
-
-    private val loggingInterceptor = HttpLoggingInterceptor().apply {
-        level = HttpLoggingInterceptor.Level.BODY
-    }
-
-    private val client = OkHttpClient.Builder()
-        .addInterceptor(loggingInterceptor)
-        .build()
-
-    @OptIn(ExperimentalSerializationApi::class)
-    val apiClient: API by lazy {
-        val retrofit = Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .client(client)
-            .addConverterFactory(
-                Json { ignoreUnknownKeys = true }
-                    .asConverterFactory("application/json".toMediaType())
-            )
-            .build()
-
-        return@lazy retrofit.create(API::class.java)
-    }
-}
