--- app/src/main/java/com/example/compose_magazin/dataAccess/models/PetCategoryData.kt
@@ -1,9 +1,0 @@
-package com.example.compose_magazin.dataAccess.models
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class PetCategoryData(
-    val id: Long,
-    val name: String? = ""
-)
